fix: change price to double

+++ android/data-transfer-objects/src/main/java/com/isranascimento/datatransferobjects/hotels/PriceResponse.kt
@@ -5,10 +5,10 @@ import com.fasterxml.jackson.annotation.JsonProperty
 data class PriceResponse(
     @JsonProperty("currency") val currency: String?,
     @JsonProperty("currency_original") val currencyOriginal: String?,
-    @JsonProperty("current_price") val currentPrice: Float?,
-    @JsonProperty("old_price") val oldPrice: Float?,
+    @JsonProperty("current_price") val currentPrice: Double?,
+    @JsonProperty("old_price") val oldPrice: Double?,
     @JsonProperty("sku") val sku: String?,
-    @JsonProperty("originalAmountPerDay") val originalAmountPerDay: Float?,
-    @JsonProperty("amountPerDay") val amountPerDay: Float?,
-    @JsonProperty("amount") val amount: Float?
+    @JsonProperty("originalAmountPerDay") val originalAmountPerDay: Double?,
+    @JsonProperty("amountPerDay") val amountPerDay: Double?,
+    @JsonProperty("amount") val amount: Double?
 )
